# chore(rxjs): update to 5.0.0-beta.2