--- a/package.json
+++ b/package.json
@@ -58,7 +58,7 @@
     "react-dom": "^0.14.7",
     "read": "^1.0.7",
     "request": "^2.69.0",
-    "rxjs": "5.0.0-beta.1",
+    "rxjs": "5.0.0-beta.2",
     "slate-irc": "karen-irc/slate-irc.git#b12d540fa231b64e4e012c9e29412a2dc448acbe",
     "socket.io": "^1.4.5",
     "socket.io-client": "^1.4.5",