https://github.com/ReactiveX/RxJS/blob/master/CHANGELOG.md#500-beta2-2016-02-10

<!-- Reviewable:start -->
[<img src="https://reviewable.io/review_button.svg" height="40" alt="Review on Reviewable"/>](https://reviewable.io/reviews/karen-irc/karen/574)
<!-- Reviewable:end -->