@saneyuki 2016-02-11
Drawback
--------------

This pull request contains serious compile performance regression :(

### Before (beta1)

Files:             111
Lines:           28560
Nodes:          119182
Identifiers:     44304
Symbols:       2960494
Types:           66178
Memory used:   165777K
I/O read:        0.03s
I/O write:       0.32s
Parse time:      0.48s
Bind time:       0.28s
Check time:      1.83s
Emit time:       0.52s
Total time:      3.10s


### After (beta1)

Files:             124
Lines:           29011
Nodes:          125219
Identifiers:     46931
Symbols:       5126214
Types:         1069572
Memory used:  1264781K
I/O read:        0.04s
I/O write:       0.31s
Parse time:      0.73s
Bind time:       0.41s
Check time:     31.48s
Emit time:       1.53s
Total time:     34.15s


I suspect this regression is caused by the complex type inference for multiple override and generics...  We may avoid this regression with annotating types explicitly, but it would be harder way, so we need to find the bottleneck.