@saneyuki 2016-02-11
> We may avoid this regression with annotating types explicitly, but it would be harder way, so we need to find

My guess was right!
I add some explicit annotation like https://github.com/saneyuki/karen/commit/c3e0926e7bb89b4682cdd47ba8778ca090d60c96, this change achieves to recover the compile performance:

```
|             | beta.1   |  beta.2  |  with annnotations |
|-------------|----------|----------|--------------------|
|Files:       |      111 |      124 |        124         |
|Lines:       |    28560 |    29011 |      29012         |
|Nodes:       |   119182 |   125219 |     125258         |
|Identifiers: |    44304 |    46931 |      46945         |
|Symbols:     |  2960494 |  5126214 |    3375603         |
|Types:       |    66178 |  1069572 |      41087         |
|Memory used: |  165777K | 1264781K |    118627K         |
|I/O read:    |    0.03s |    0.04s |      0.04s         |
|I/O write:   |    0.32s |    0.31s |      0.22s         |
|Parse time:  |    0.48s |    0.73s |      0.64s         |
|Bind time:   |    0.28s |    0.41s |      0.31s         |
|Check time:  |    1.83s |   31.48s |      1.67s         |
|Emit time:   |    0.52s |    1.53s |      0.43s         |
|Total time:  |    3.10s |   34.15s |      3.05s         |
```

By this change, we found a some points:

- TypeScript compiler is not good performance to infer types if a functions has a very complex signature with multiple overloading and type parameters for generics.
   - [`combineLatest()`](https://github.com/ReactiveX/RxJS/blob/603c9eb1e0cb97f0904734e594dce398776306bb/src/operator/combineLatest.ts#L42-L87)
   - [`zip()`](https://github.com/ReactiveX/RxJS/blob/603c9eb1e0cb97f0904734e594dce398776306bb/src/operator/zip.ts#L18-L53)
   - and others...
- The current TypeScript compiler try to resolve such complex definitions without give up to infer types as a compile error.
   - In this case, TypeScript may enter in a edge case....
- To avoid this perf problem, we need write a types for their complex functions or the moment.

We should report this as a issue to TypeScript or RxJS, but it's difficult to describe a problem...