@dokidokivisual 2016-02-11
:pushpin: Commit c3e0926 has been approved by `saneyuki`

<!-- @dokidokivisual r=saneyuki c3e0926e7bb89b4682cdd47ba8778ca090d60c96 -->