@dokidokivisual 2016-02-11
:sunny: Test successful - [travis](https://travis-ci.org/karen-irc/karen/builds/108571039)