@saneyuki 2016-02-16
Filed as the problem: https://github.com/Microsoft/TypeScript/issues/7082